+++ services/ims/v2/model/CreateWholeImageRequestBody.ts
@@ -13,8 +13,7 @@ export class CreateWholeImageRequestBody {
     private 'max_ram'?: number;
     private 'min_ram'?: number;
     private 'vault_id'?: string;
-    public constructor(instanceId?: string, name?: string) { 
-        this['instance_id'] = instanceId;
+    public constructor(name?: string) { 
         this['name'] = name;
     }
     public withDescription(description: string): CreateWholeImageRequestBody {