+++ services/rds/v3/model/CreateJobScheduleRequest.ts
@@ -1,0 +1,12 @@
+import { CreateJobScheduleRequestBody } from './CreateJobScheduleRequestBody';
+
+
+export class CreateJobScheduleRequest {
+    public body?: CreateJobScheduleRequestBody;
+    public constructor() { 
+    }
+    public withBody(body: CreateJobScheduleRequestBody): CreateJobScheduleRequest {
+        this['body'] = body;
+        return this;
+    }
+}