+++ services/rds/v3/model/ModifyJobScheduleResponse.ts
@@ -1,12 +1,12 @@
 
 import { SdkResponse } from "@huaweicloud/huaweicloud-sdk-core/SdkResponse";
 
-export class SwitchDbAgentJobResponse extends SdkResponse {
+export class ModifyJobScheduleResponse extends SdkResponse {
     public resp?: string;
     public constructor() { 
         super();
     }
-    public withResp(resp: string): SwitchDbAgentJobResponse {
+    public withResp(resp: string): ModifyJobScheduleResponse {
         this['resp'] = resp;
         return this;
     }