+++ services/rds/v3/model/ValidateInstanceConnectionResponse.ts
@@ -2,7 +2,12 @@
 import { SdkResponse } from "@huaweicloud/huaweicloud-sdk-core/SdkResponse";
 
 export class ValidateInstanceConnectionResponse extends SdkResponse {
+    public resp?: string;
     public constructor() { 
         super();
+    }
+    public withResp(resp: string): ValidateInstanceConnectionResponse {
+        this['resp'] = resp;
+        return this;
     }
 }